[DEV-2096] rename bitcoinbased
