+++ src/subdomains/supporting/payout/strategies/payout/impl/bitcoin.strategy.ts
@@ -8,12 +8,12 @@ import { NotificationService } from 'src/subdomains/supporting/notification/serv
 import { PayoutOrder, PayoutOrderContext } from '../../../entities/payout-order.entity';
 import { FeeResult } from '../../../interfaces';
 import { PayoutOrderRepository } from '../../../repositories/payout-order.repository';
-import { PayoutGroup } from '../../../services/base/payout-bitcoinbased.service';
+import { PayoutGroup } from '../../../services/base/payout-bitcoin-based.service';
 import { PayoutBitcoinService } from '../../../services/payout-bitcoin.service';
-import { BitcoinbasedStrategy } from './base/bitcoinbased.strategy';
+import { BitcoinBasedStrategy } from './base/bitcoin-based.strategy';
 
 @Injectable()
-export class BitcoinStrategy extends BitcoinbasedStrategy {
+export class BitcoinStrategy extends BitcoinBasedStrategy {
   protected readonly logger = new DfxLogger(BitcoinStrategy);
 
   private readonly averageTransactionSize = 180; // vBytes
